feat: add a feature flag for count alarms
Adds a feature flag to turn off alarms that have a tendency to go off
during the collection of debug metrics.

The reason for turning this off versus just increasing metric count is
that we have so far been pretty bad about being able to predict traffic
patterns with debug metrics turned on.

We should be able to determine failure of the metrics collection service
through other means such as an increase in errors or a drop off of
invocations.